requirements.txt
numpy<2.0.0
moviepy==1.0.3
scenedetect==0.6.2
opencv-python==4.8.1.78
torch>=1.7.1
clip @ git+https://github.com/openai/CLIP.git
Pillow>=8.0.0
matplotlib
flow_vis
cotracker @ git+https://github.com/facebookresearch/co-tracker.git
wget 
scikit-learn